requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
firebase-admin==6.5.0
pydantic==2.8.2
qrcode==7.4.2
Pillow==10.4.0
shortuuid==1.0.13
